requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
emoji==2.0.0
frozenlist==1.3.0
idna==3.3
keyboard==0.13.5
multidict==6.0.2
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
urllib3==1.26.10
yarl==1.7.2
